refactor: Extract TransferAmount from modals
